backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & HTTP
aiohttp==3.9.1
httpx==0.26.0
asyncio==3.4.3

# ML & AI
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
spacy==3.7.2
scikit-learn==1.4.0
lightgbm==4.2.0
numpy==1.26.3
pandas==2.1.4

# Free AI APIs
huggingface-hub==0.20.2
groq==0.4.1
google-generativeai==0.3.2

# Database (SQLite for local, PostgreSQL for production)
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.13.1

# Database & Storage
boto3==1.34.34
aioboto3==12.3.0
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# Validation & Data Processing
email-validator==2.1.0
phonenumbers==8.13.27
python-dateutil==2.8.2
pycountry==23.12.11

# Monitoring & Logging
python-json-logger==2.0.7
aws-xray-sdk==2.12.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
orjson==3.9.12
